+++ LinearAlgebra/include/MatrixHelpers.hpp
@@ -19,7 +19,7 @@ namespace LinearAlgebra::Matrix
 
     //  Search the index of the first row with a non-zero entry below the current element (that is, in the same column).
     //  Used internally on the upper matrix which is always double, therefore only double spec.
-    std::optional<unsigned int> findNonZeroPivot(const Matrix<double>& mat, unsigned int, unsigned int);
+    std::optional<unsigned int> findNonZeroPivot(const Matrix<long double>& mat, unsigned int, unsigned int);
 
     template <typename T>
     bool isLowerTriangular(const Matrix<T>& m);
